Improvements to result handling in more complex scenarios
Added some tests for managing flows of results, where getting a result from one screen opens another screen, and cancelling from an inner screen cancels the outer screen. Also fixed a bug with primary navigation fragment, so that this is only changed if the fragment closing is already the primary navigation fragment

Changed navigation handle execution to use lifecycle scopes and launchWhenCreated for more consistent result handling.

Added extension for sending a result from an executor, which is useful for writing custom overrides.